feat: button이 있는 경우와 없는 경우의 렌더링을 tooltipContent에서 분리
